+++ cartridge/confapplier.lua
@@ -28,6 +28,7 @@ local remote_control = require('cartridge.remote-control')
 local cluster_cookie = require('cartridge.cluster-cookie')
 local ClusterwideConfig = require('cartridge.clusterwide-config')
 local logging_whitelist = require('cartridge.logging_whitelist')
+local invalid_format = require('cartridge.invalid-format')
 
 yaml.cfg({
     encode_load_metatables = false,
@@ -542,11 +543,12 @@ local function boot_instance(clusterwide_config)
     -- This operation may be long
     -- It recovers snapshot
     -- Or bootstraps replication
+    invalid_format.start_check()
     local snap1 = hotreload.snap_fibers()
     box.cfg(box_opts)
     local snap2 = hotreload.snap_fibers()
     hotreload.whitelist_fibers(hotreload.diff(snap1, snap2))
-
+    invalid_format.end_check()
     require('membership.options').SUSPICIOUSNESS = true
 
     local username = cluster_cookie.username()